Begin changing to the whiteplain theme
Install the theme in the themes directory, start making the home page an
about page, and update config file for the new theme. At this point, I
decided it didn't even look that much better than what I'd already been
using with the theme I'd made myself from bootstrap-minimal. So I just
want to save this work and move on.