Add mainSections site param
Add a mainSections parameter to the site params in the config file, so
that only content of type "posts" is shown in the list of recent posts.
After adding the content type "page", only content of type page was
shown in the list of recent posts, because mainSections defaults to the
section with the most pages.
https://gohugo.io/functions/where/#mainsections